+++ tests/api-testing/helpers/workflow/delete_objects.py
