+++ tests/api-testing/helpers/workflow/verifydeleted_objects.py
