+++ tests/api-testing/tests/test_workflow.py
@@ -1,11 +1,10 @@
 import os
-from create_objects import create_objects
-
-from verify_objects import verify_objects
-from edit_objects import edit_objects
-from delete_objects import delete_objects
-from verifydeleted_objects import verify_deleted_objects       
-from verifyedited_objects import verify_edited_objects
+from helpers.workflow.create_objects import create_objects
+from helpers.workflow.verify_objects import verify_objects
+from helpers.workflow.edit_objects import edit_objects
+from helpers.workflow.delete_objects import delete_objects
+from helpers.workflow.verifydeleted_objects import verify_deleted_objects
+from helpers.workflow.verifyedited_objects import verify_edited_objects
